chore: replaced species_aligner_analysis.py with the new version
